REFACTOR format to pylint FIX pylint workflow #4
